Make the async Client default
The previously default Client is moved to `reqwest::blocking`, while the
async client becomes the main API.

Closes #622
