+++ examples/form.rs
@@ -1,7 +1,9 @@
-fn main() {
+#[tokio::main]
+async fn main() {
     reqwest::Client::new()
         .post("http://www.baidu.com")
         .form(&[("one", "1")])
         .send()
+        .await
         .unwrap();
 }
