+++ tests/client.rs
@@ -1,8 +1,30 @@
 #[macro_use]
 mod support;
 
-#[test]
-fn test_response_text() {
+use std::io::Write;
+use std::time::Duration;
+
+use futures::TryStreamExt;
+
+use reqwest::multipart::{Form, Part};
+use reqwest::{Body, Client};
+
+use bytes::Bytes;
+
+#[tokio::test]
+async fn gzip_response() {
+    gzip_case(10_000, 4096).await;
+}
+
+#[tokio::test]
+async fn gzip_single_byte_chunks() {
+    gzip_case(10, 1).await;
+}
+
+#[tokio::test]
+async fn response_text() {
+    let _ = env_logger::try_init();
+
     let server = server! {
         request: b"\
             GET /text HTTP/1.1\r\n\
@@ -14,65 +36,27 @@ fn test_response_text() {
             ",
         response: b"\
             HTTP/1.1 200 OK\r\n\
-            Server: test\r\n\
             Content-Length: 5\r\n\
             \r\n\
             Hello\
             "
     };
 
-    let url = format!("http://{}/text", server.addr());
-    let mut res = reqwest::get(&url).unwrap();
-    assert_eq!(res.url().as_str(), &url);
-    assert_eq!(res.status(), reqwest::StatusCode::OK);
-    assert_eq!(res.headers().get(reqwest::header::SERVER).unwrap(), &"test");
-    assert_eq!(
-        res.headers().get(reqwest::header::CONTENT_LENGTH).unwrap(),
-        &"5"
-    );
+    let client = Client::new();
 
-    let body = res.text().unwrap();
-    assert_eq!(b"Hello", body.as_bytes());
+    let mut res = client
+        .get(&format!("http://{}/text", server.addr()))
+        .send()
+        .await
+        .expect("Failed to get");
+    let text = res.text().await.expect("Failed to get text");
+    assert_eq!("Hello", text);
 }
 
-#[test]
-fn test_response_non_utf_8_text() {
-    let server = server! {
-        request: b"\
-            GET /text HTTP/1.1\r\n\
-            user-agent: $USERAGENT\r\n\
-            accept: */*\r\n\
-            accept-encoding: gzip\r\n\
-            host: $HOST\r\n\
-            \r\n\
-            ",
-        response: b"\
-            HTTP/1.1 200 OK\r\n\
-            Server: test\r\n\
-            Content-Length: 4\r\n\
-            Content-Type: text/plain; charset=gbk\r\n\
-            \r\n\
-            \xc4\xe3\xba\xc3\
-            "
-    };
+#[tokio::test]
+async fn response_json() {
+    let _ = env_logger::try_init();
 
-    let url = format!("http://{}/text", server.addr());
-    let mut res = reqwest::get(&url).unwrap();
-    assert_eq!(res.url().as_str(), &url);
-    assert_eq!(res.status(), reqwest::StatusCode::OK);
-    assert_eq!(res.headers().get(reqwest::header::SERVER).unwrap(), &"test");
-    assert_eq!(
-        res.headers().get(reqwest::header::CONTENT_LENGTH).unwrap(),
-        &"4"
-    );
-
-    let body = res.text().unwrap();
-    assert_eq!("你好", &body);
-    assert_eq!(b"\xe4\xbd\xa0\xe5\xa5\xbd", body.as_bytes()); // Now it's utf-8
-}
-
-#[test]
-fn test_response_json() {
     let server = server! {
         request: b"\
             GET /json HTTP/1.1\r\n\
@@ -84,32 +68,100 @@ fn test_response_json() {
             ",
         response: b"\
             HTTP/1.1 200 OK\r\n\
-            Server: test\r\n\
             Content-Length: 7\r\n\
             \r\n\
             \"Hello\"\
             "
     };
 
-    let url = format!("http://{}/json", server.addr());
-    let mut res = reqwest::get(&url).unwrap();
-    assert_eq!(res.url().as_str(), &url);
-    assert_eq!(res.status(), reqwest::StatusCode::OK);
-    assert_eq!(res.headers().get(reqwest::header::SERVER).unwrap(), &"test");
-    assert_eq!(
-        res.headers().get(reqwest::header::CONTENT_LENGTH).unwrap(),
-        &"7"
-    );
+    let client = Client::new();
 
-    let body = res.json::<String>().unwrap();
-    assert_eq!("Hello", body);
+    let mut res = client
+        .get(&format!("http://{}/json", server.addr()))
+        .send()
+        .await
+        .expect("Failed to get");
+    let text = res.json::<String>().await.expect("Failed to get json");
+    assert_eq!("Hello", text);
 }
 
-#[test]
-fn test_response_copy_to() {
+#[tokio::test]
+async fn multipart() {
+    let _ = env_logger::try_init();
+
+    let stream = futures::stream::once(futures::future::ready::<Result<_, hyper::Error>>(Ok(
+        hyper::Chunk::from("part1 part2".to_owned()),
+    )));
+    let part = Part::stream(stream);
+
+    let form = Form::new().text("foo", "bar").part("part_stream", part);
+
+    let expected_body = format!(
+        "\
+         24\r\n\
+         --{0}\r\n\r\n\
+         2E\r\n\
+         Content-Disposition: form-data; name=\"foo\"\r\n\r\n\r\n\
+         3\r\n\
+         bar\r\n\
+         2\r\n\
+         \r\n\r\n\
+         24\r\n\
+         --{0}\r\n\r\n\
+         36\r\n\
+         Content-Disposition: form-data; name=\"part_stream\"\r\n\r\n\r\n\
+         B\r\n\
+         part1 part2\r\n\
+         2\r\n\
+         \r\n\r\n\
+         26\r\n\
+         --{0}--\r\n\r\n\
+         0\r\n\r\n\
+         ",
+        form.boundary()
+    );
+
+    let server = server! {
+        request: format!("\
+            POST /multipart/1 HTTP/1.1\r\n\
+            content-type: multipart/form-data; boundary={}\r\n\
+            user-agent: $USERAGENT\r\n\
+            accept: */*\r\n\
+            accept-encoding: gzip\r\n\
+            host: $HOST\r\n\
+            transfer-encoding: chunked\r\n\
+            \r\n\
+            {}\
+            ", form.boundary(), expected_body),
+        response: b"\
+            HTTP/1.1 200 OK\r\n\
+            Server: multipart\r\n\
+            Content-Length: 0\r\n\
+            \r\n\
+            "
+    };
+
+    let url = format!("http://{}/multipart/1", server.addr());
+
+    let client = Client::new();
+
+    let res = client
+        .post(&url)
+        .multipart(form)
+        .send()
+        .await
+        .expect("Failed to post multipart");
+    assert_eq!(res.url().as_str(), &url);
+    assert_eq!(res.status(), reqwest::StatusCode::OK);
+}
+
+#[tokio::test]
+async fn request_timeout() {
+    let _ = env_logger::try_init();
+
     let server = server! {
         request: b"\
-            GET /1 HTTP/1.1\r\n\
+            GET /slow HTTP/1.1\r\n\
             user-agent: $USERAGENT\r\n\
             accept: */*\r\n\
             accept-encoding: gzip\r\n\
@@ -118,31 +170,35 @@ fn test_response_copy_to() {
             ",
         response: b"\
             HTTP/1.1 200 OK\r\n\
-            Server: test\r\n\
             Content-Length: 5\r\n\
             \r\n\
             Hello\
-            "
+            ",
+        read_timeout: Duration::from_secs(2)
     };
 
-    let url = format!("http://{}/1", server.addr());
-    let mut res = reqwest::get(&url).unwrap();
-    assert_eq!(res.url().as_str(), &url);
-    assert_eq!(res.status(), reqwest::StatusCode::OK);
-    assert_eq!(res.headers().get(reqwest::header::SERVER).unwrap(), &"test");
-    assert_eq!(
-        res.headers().get(reqwest::header::CONTENT_LENGTH).unwrap(),
-        &"5"
-    );
+    let client = Client::builder()
+        .timeout(Duration::from_millis(500))
+        .build()
+        .unwrap();
 
-    assert_eq!("Hello".to_owned(), res.text().unwrap());
+    let url = format!("http://{}/slow", server.addr());
+
+    let res = client.get(&url).send().await;
+
+    let err = res.unwrap_err();
+
+    assert!(err.is_timeout());
+    assert_eq!(err.url().map(|u| u.as_str()), Some(url.as_str()));
 }
 
-#[test]
-fn test_get() {
+#[tokio::test]
+async fn response_timeout() {
+    let _ = env_logger::try_init();
+
     let server = server! {
         request: b"\
-            GET /1 HTTP/1.1\r\n\
+            GET /slow HTTP/1.1\r\n\
             user-agent: $USERAGENT\r\n\
             accept: */*\r\n\
             accept-encoding: gzip\r\n\
@@ -151,362 +207,119 @@ fn test_get() {
             ",
         response: b"\
             HTTP/1.1 200 OK\r\n\
-            Server: test\r\n\
-            Content-Length: 0\r\n\
-            \r\n\
-            "
-    };
-
-    let url = format!("http://{}/1", server.addr());
-    let mut res = reqwest::get(&url).unwrap();
-
-    assert_eq!(res.url().as_str(), &url);
-    assert_eq!(res.status(), reqwest::StatusCode::OK);
-    assert_eq!(res.headers().get(reqwest::header::SERVER).unwrap(), &"test");
-    assert_eq!(
-        res.headers().get(reqwest::header::CONTENT_LENGTH).unwrap(),
-        &"0"
-    );
-    assert_eq!(res.remote_addr(), Some(server.addr()));
-
-    assert_eq!(res.text().unwrap().len(), 0)
-}
-
-#[test]
-fn test_post() {
-    let server = server! {
-        request: b"\
-            POST /2 HTTP/1.1\r\n\
-            content-length: 5\r\n\
-            user-agent: $USERAGENT\r\n\
-            accept: */*\r\n\
-            accept-encoding: gzip\r\n\
-            host: $HOST\r\n\
+            Content-Length: 5\r\n\
             \r\n\
             Hello\
             ",
+        write_timeout: Duration::from_secs(2)
+    };
+
+    let client = Client::builder()
+        .timeout(Duration::from_millis(500))
+        .build()
+        .unwrap();
+
+    let url = format!("http://{}/slow", server.addr());
+    let res = client.get(&url).send().await.expect("Failed to get");
+    let body: Result<_, _> = res.into_body().try_concat().await;
+
+    let err = body.unwrap_err();
+
+    assert!(err.is_timeout());
+}
+
+async fn gzip_case(response_size: usize, chunk_size: usize) {
+    let content: String = (0..response_size)
+        .into_iter()
+        .map(|i| format!("test {}", i))
+        .collect();
+    let mut encoder = libflate::gzip::Encoder::new(Vec::new()).unwrap();
+    match encoder.write(content.as_bytes()) {
+        Ok(n) => assert!(n > 0, "Failed to write to encoder."),
+        _ => panic!("Failed to gzip encode string."),
+    };
+
+    let gzipped_content = encoder.finish().into_result().unwrap();
+
+    let mut response = format!(
+        "\
+         HTTP/1.1 200 OK\r\n\
+         Server: test-accept\r\n\
+         Content-Encoding: gzip\r\n\
+         Content-Length: {}\r\n\
+         \r\n",
+        &gzipped_content.len()
+    )
+    .into_bytes();
+    response.extend(&gzipped_content);
+
+    let server = server! {
+        request: b"\
+            GET /gzip HTTP/1.1\r\n\
+            user-agent: $USERAGENT\r\n\
+            accept: */*\r\n\
+            accept-encoding: gzip\r\n\
+            host: $HOST\r\n\
+            \r\n\
+            ",
+        chunk_size: chunk_size,
+        write_timeout: Duration::from_millis(10),
+        response: response
+    };
+
+    let client = Client::new();
+
+    let mut res = client
+        .get(&format!("http://{}/gzip", server.addr()))
+        .send()
+        .await
+        .expect("response");
+
+    let body = res.text().await.expect("text");
+    assert_eq!(body, content);
+}
+
+#[tokio::test]
+async fn body_stream() {
+    let _ = env_logger::try_init();
+
+    let source = futures::stream::iter::<Vec<Result<Bytes, std::io::Error>>>(vec![
+        Ok(Bytes::from_static(b"123")),
+        Ok(Bytes::from_static(b"4567")),
+    ]);
+
+    let expected_body = "3\r\n123\r\n4\r\n4567\r\n0\r\n\r\n";
+
+    let server = server! {
+        request: format!("\
+            POST /post HTTP/1.1\r\n\
+            user-agent: $USERAGENT\r\n\
+            accept: */*\r\n\
+            accept-encoding: gzip\r\n\
+            host: $HOST\r\n\
+            transfer-encoding: chunked\r\n\
+            \r\n\
+            {}\
+            ", expected_body),
         response: b"\
             HTTP/1.1 200 OK\r\n\
             Server: post\r\n\
-            Content-Length: 0\r\n\
+            Content-Length: 7\r\n\
             \r\n\
             "
     };
 
-    let url = format!("http://{}/2", server.addr());
-    let mut res = reqwest::Client::new()
+    let url = format!("http://{}/post", server.addr());
+
+    let client = Client::new();
+
+    let res = client
         .post(&url)
-        .body("Hello")
+        .body(Body::wrap_stream(source))
         .send()
-        .unwrap();
+        .await
+        .expect("Failed to post");
 
     assert_eq!(res.url().as_str(), &url);
     assert_eq!(res.status(), reqwest::StatusCode::OK);
-    assert_eq!(res.headers().get(reqwest::header::SERVER).unwrap(), &"post");
-    assert_eq!(
-        res.headers().get(reqwest::header::CONTENT_LENGTH).unwrap(),
-        &"0"
-    );
-
-    assert_eq!(res.text().unwrap().len(), 0)
-}
-
-#[test]
-fn test_post_form() {
-    let server = server! {
-        request: b"\
-            POST /form HTTP/1.1\r\n\
-            content-type: application/x-www-form-urlencoded\r\n\
-            content-length: 24\r\n\
-            user-agent: $USERAGENT\r\n\
-            accept: */*\r\n\
-            accept-encoding: gzip\r\n\
-            host: $HOST\r\n\
-            \r\n\
-            hello=world&sean=monstar\
-            ",
-        response: b"\
-            HTTP/1.1 200 OK\r\n\
-            Server: post-form\r\n\
-            Content-Length: 0\r\n\
-            \r\n\
-            "
-    };
-
-    let form = &[("hello", "world"), ("sean", "monstar")];
-
-    let url = format!("http://{}/form", server.addr());
-    let res = reqwest::Client::new()
-        .post(&url)
-        .form(form)
-        .send()
-        .expect("request send");
-
-    assert_eq!(res.url().as_str(), &url);
-    assert_eq!(res.status(), reqwest::StatusCode::OK);
-}
-
-/// Calling `Response::error_for_status`` on a response with status in 4xx
-/// returns a error.
-#[test]
-fn test_error_for_status_4xx() {
-    let server = server! {
-        request: b"\
-            GET /1 HTTP/1.1\r\n\
-            user-agent: $USERAGENT\r\n\
-            accept: */*\r\n\
-            accept-encoding: gzip\r\n\
-            host: $HOST\r\n\
-            \r\n\
-            ",
-        response: b"\
-            HTTP/1.1 400 OK\r\n\
-            Server: test\r\n\
-            Content-Length: 0\r\n\
-            \r\n\
-            "
-    };
-
-    let url = format!("http://{}/1", server.addr());
-    let res = reqwest::get(&url).unwrap();
-
-    let err = res.error_for_status().err().unwrap();
-    assert!(err.is_client_error());
-    assert_eq!(err.status(), Some(reqwest::StatusCode::BAD_REQUEST));
-}
-
-/// Calling `Response::error_for_status`` on a response with status in 5xx
-/// returns a error.
-#[test]
-fn test_error_for_status_5xx() {
-    let server = server! {
-        request: b"\
-            GET /1 HTTP/1.1\r\n\
-            user-agent: $USERAGENT\r\n\
-            accept: */*\r\n\
-            accept-encoding: gzip\r\n\
-            host: $HOST\r\n\
-            \r\n\
-            ",
-        response: b"\
-            HTTP/1.1 500 OK\r\n\
-            Server: test\r\n\
-            Content-Length: 0\r\n\
-            \r\n\
-            "
-    };
-
-    let url = format!("http://{}/1", server.addr());
-    let res = reqwest::get(&url).unwrap();
-
-    let err = res.error_for_status().err().unwrap();
-    assert!(err.is_server_error());
-    assert_eq!(
-        err.status(),
-        Some(reqwest::StatusCode::INTERNAL_SERVER_ERROR)
-    );
-}
-
-#[test]
-fn test_default_headers() {
-    use reqwest::header;
-    let mut headers = header::HeaderMap::with_capacity(1);
-    headers.insert(header::COOKIE, header::HeaderValue::from_static("a=b;c=d"));
-    let client = reqwest::Client::builder()
-        .default_headers(headers)
-        .build()
-        .unwrap();
-
-    let server = server! {
-        request: b"\
-            GET /1 HTTP/1.1\r\n\
-            user-agent: $USERAGENT\r\n\
-            accept: */*\r\n\
-            cookie: a=b;c=d\r\n\
-            accept-encoding: gzip\r\n\
-            host: $HOST\r\n\
-            \r\n\
-            ",
-        response: b"\
-            HTTP/1.1 200 OK\r\n\
-            Server: test\r\n\
-            Content-Length: 0\r\n\
-            \r\n\
-            "
-    };
-
-    let url = format!("http://{}/1", server.addr());
-    let res = client.get(&url).send().unwrap();
-
-    assert_eq!(res.url().as_str(), &url);
-    assert_eq!(res.status(), reqwest::StatusCode::OK);
-    assert_eq!(res.headers().get(reqwest::header::SERVER).unwrap(), &"test");
-    assert_eq!(
-        res.headers().get(reqwest::header::CONTENT_LENGTH).unwrap(),
-        &"0"
-    );
-
-    let server = server! {
-        request: b"\
-            GET /2 HTTP/1.1\r\n\
-            user-agent: $USERAGENT\r\n\
-            accept: */*\r\n\
-            cookie: a=b;c=d\r\n\
-            accept-encoding: gzip\r\n\
-            host: $HOST\r\n\
-            \r\n\
-            ",
-        response: b"\
-            HTTP/1.1 200 OK\r\n\
-            Server: test\r\n\
-            Content-Length: 0\r\n\
-            \r\n\
-            "
-    };
-
-    let url = format!("http://{}/2", server.addr());
-    let res = client.get(&url).send().unwrap();
-
-    assert_eq!(res.url().as_str(), &url);
-    assert_eq!(res.status(), reqwest::StatusCode::OK);
-    assert_eq!(res.headers().get(reqwest::header::SERVER).unwrap(), &"test");
-    assert_eq!(
-        res.headers().get(reqwest::header::CONTENT_LENGTH).unwrap(),
-        &"0"
-    );
-}
-
-#[test]
-fn test_override_default_headers() {
-    use reqwest::header;
-    let mut headers = header::HeaderMap::with_capacity(1);
-    headers.insert(
-        header::AUTHORIZATION,
-        header::HeaderValue::from_static("iamatoken"),
-    );
-    let client = reqwest::Client::builder()
-        .default_headers(headers)
-        .build()
-        .unwrap();
-
-    let server = server! {
-        request: b"\
-            GET /3 HTTP/1.1\r\n\
-            authorization: secret\r\n\
-            user-agent: $USERAGENT\r\n\
-            accept: */*\r\n\
-            accept-encoding: gzip\r\n\
-            host: $HOST\r\n\
-            \r\n\
-            ",
-        response: b"\
-            HTTP/1.1 200 OK\r\n\
-            Server: test\r\n\
-            Content-Length: 0\r\n\
-            \r\n\
-            "
-    };
-
-    let url = format!("http://{}/3", server.addr());
-    let res = client
-        .get(&url)
-        .header(
-            header::AUTHORIZATION,
-            header::HeaderValue::from_static("secret"),
-        )
-        .send()
-        .unwrap();
-
-    assert_eq!(res.url().as_str(), &url);
-    assert_eq!(res.status(), reqwest::StatusCode::OK);
-    assert_eq!(res.headers().get(reqwest::header::SERVER).unwrap(), &"test");
-    assert_eq!(
-        res.headers().get(reqwest::header::CONTENT_LENGTH).unwrap(),
-        &"0"
-    );
-}
-
-#[test]
-fn test_appended_headers_not_overwritten() {
-    let client = reqwest::Client::new();
-
-    let server = server! {
-        request: b"\
-            GET /4 HTTP/1.1\r\n\
-            accept: application/json\r\n\
-            accept: application/json+hal\r\n\
-            user-agent: $USERAGENT\r\n\
-            accept-encoding: gzip\r\n\
-            host: $HOST\r\n\
-            \r\n\
-            ",
-        response: b"\
-            HTTP/1.1 200 OK\r\n\
-            Server: test\r\n\
-            Content-Length: 0\r\n\
-            \r\n\
-            "
-    };
-
-    let url = format!("http://{}/4", server.addr());
-    let res = client
-        .get(&url)
-        .header(header::ACCEPT, "application/json")
-        .header(header::ACCEPT, "application/json+hal")
-        .send()
-        .unwrap();
-
-    assert_eq!(res.url().as_str(), &url);
-    assert_eq!(res.status(), reqwest::StatusCode::OK);
-    assert_eq!(res.headers().get(reqwest::header::SERVER).unwrap(), &"test");
-    assert_eq!(
-        res.headers().get(reqwest::header::CONTENT_LENGTH).unwrap(),
-        &"0"
-    );
-
-    // make sure this also works with default headers
-    use reqwest::header;
-    let mut headers = header::HeaderMap::with_capacity(1);
-    headers.insert(
-        header::ACCEPT,
-        header::HeaderValue::from_static("text/html"),
-    );
-    let client = reqwest::Client::builder()
-        .default_headers(headers)
-        .build()
-        .unwrap();
-
-    let server = server! {
-        request: b"\
-            GET /4 HTTP/1.1\r\n\
-            accept: application/json\r\n\
-            accept: application/json+hal\r\n\
-            user-agent: $USERAGENT\r\n\
-            accept-encoding: gzip\r\n\
-            host: $HOST\r\n\
-            \r\n\
-            ",
-        response: b"\
-            HTTP/1.1 200 OK\r\n\
-            Server: test\r\n\
-            Content-Length: 0\r\n\
-            \r\n\
-            "
-    };
-
-    let url = format!("http://{}/4", server.addr());
-    let res = client
-        .get(&url)
-        .header(header::ACCEPT, "application/json")
-        .header(header::ACCEPT, "application/json+hal")
-        .send()
-        .unwrap();
-
-    assert_eq!(res.url().as_str(), &url);
-    assert_eq!(res.status(), reqwest::StatusCode::OK);
-    assert_eq!(res.headers().get(reqwest::header::SERVER).unwrap(), &"test");
-    assert_eq!(
-        res.headers().get(reqwest::header::CONTENT_LENGTH).unwrap(),
-        &"0"
-    );
 }
